Add initial implementation of device property functionality for the SG380 series signal generators.
This code adds an output enable widget and a modulation control widget to the BLACS tab. The type of modulation is set at the connectiontable. Everything else about the modulation is controlled from BLACS.

At this point, device properties are not remote value checked and there is no control of the device parameters from the connectiontable.